Add more integration in scene
