+++ manim/renderer/opengl_renderer.py
@@ -1,6 +1,5 @@
 import itertools as it
 import time
-from queue import Queue
 
 import moderngl
 import numpy as np
@@ -45,15 +44,14 @@ class OpenGLCamera(OpenGLMobject):
         minimum_polar_angle=-PI / 2,
         maximum_polar_angle=PI / 2,
         model_matrix=None,
-            scene=None,
         **kwargs,
     ):
-        self.scene = scene # todo remove
         self.use_z_index = True
         self.frame_rate = 60
         self.orthographic = orthographic
         self.minimum_polar_angle = minimum_polar_angle
         self.maximum_polar_angle = maximum_polar_angle
+
         if self.orthographic:
             self.projection_matrix = opengl.orthographic_projection_matrix()
             self.unformatted_projection_matrix = opengl.orthographic_projection_matrix(
@@ -232,7 +230,11 @@ class OpenGLRenderer:
         self.num_plays = 0
         self.interactive_mode = False
         self.widgets = [] if not widgets else widgets
+        self.mouse_press_callbacks = []
+        self.pressed_keys = set()
+        self.key_to_function_map = {}
 
+        self.camera_target = ORIGIN
         self.camera = OpenGLCamera()
 
         # Initialize texture map.
@@ -241,11 +243,12 @@ class OpenGLRenderer:
         self._background_color = color_to_rgba(config["background_color"], 1.0)
 
     def init_scene(self, scene=None):
-        self.scene = scene # todo remove
         if not hasattr(self, "window"):
             if self.should_create_window():
                 from .opengl_renderer_window import Window
 
+                self.mouse_point = OpenGLPoint()
+                self.mouse_drag_point = OpenGLPoint()
                 self.window = Window(self)
                 self.context = self.window.ctx
                 self.frame_buffer_object = self.context.detect_framebuffer()
@@ -405,6 +408,9 @@ class OpenGLRenderer:
     def before_animation(self):
         self.animation_start_time = time.time()
 
+    def has_interaction(self):
+        return True
+
     def render(
         self,
         frame_offset,
@@ -543,3 +549,140 @@ class OpenGLRenderer:
     @background_color.setter
     def background_color(self, value):
         self._background_color = color_to_rgba(value, 1.0)
+
+    def on_mouse_motion(self, point, d_point):
+        self.mouse_point.move_to(point)
+        if SHIFT_VALUE in self.pressed_keys:
+            shift = -d_point
+            shift[0] *= self.camera.get_width() / 2
+            shift[1] *= self.camera.get_height() / 2
+            transform = self.camera.inverse_rotation_matrix
+            shift = np.dot(np.transpose(transform), shift)
+            self.camera.shift(shift)
+
+    def on_mouse_scroll(self, point, offset):
+        if not config.use_projection_stroke_shaders:
+            factor = 1 + np.arctan(-2.1 * offset[1])
+            self.camera.scale(factor, about_point=self.camera_target)
+        self.mouse_scroll_orbit_controls(point, offset)
+
+    def on_key_press(self, symbol, modifiers):
+        try:
+            char = chr(symbol)
+        except OverflowError:
+            logger.warning("The value of the pressed key is too large.")
+            return
+
+        if char == "r":
+            self.camera.to_default_state()
+            self.camera_target = np.array([0, 0, 0], dtype=np.float32)
+        elif char == "q":
+            self.quit_interaction = True
+        else:
+            if char in self.key_to_function_map:
+                self.key_to_function_map[char]()
+
+    def on_key_release(self, symbol, modifiers):
+        pass
+
+    def on_mouse_drag(self, point, d_point, buttons, modifiers):
+        self.mouse_drag_point.move_to(point)
+        if buttons == 1:
+            self.camera.increment_theta(-d_point[0])
+            self.camera.increment_phi(d_point[1])
+        elif buttons == 4:
+            camera_x_axis = self.camera.model_matrix[:3, 0]
+            horizontal_shift_vector = -d_point[0] * camera_x_axis
+            vertical_shift_vector = -d_point[1] * np.cross(OUT, camera_x_axis)
+            total_shift_vector = horizontal_shift_vector + vertical_shift_vector
+            self.camera.shift(1.1 * total_shift_vector)
+
+        self.mouse_drag_orbit_controls(point, d_point, buttons, modifiers)
+
+    def mouse_scroll_orbit_controls(self, point, offset):
+        camera_to_target = self.camera_target - self.camera.get_position()
+        camera_to_target *= np.sign(offset[1])
+        shift_vector = 0.01 * camera_to_target
+        self.camera.model_matrix = (
+            opengl.translation_matrix(*shift_vector) @ self.camera.model_matrix
+        )
+
+    def mouse_drag_orbit_controls(self, point, d_point, buttons, modifiers):
+        # Left click drag.
+        if buttons == 1:
+            # Translate to target the origin and rotate around the z axis.
+            self.camera.model_matrix = (
+                opengl.rotation_matrix(z=-d_point[0])
+                @ opengl.translation_matrix(*-self.camera_target)
+                @ self.camera.model_matrix
+            )
+
+            # Rotation off of the z axis.
+            camera_position = self.camera.get_position()
+            camera_y_axis = self.camera.model_matrix[:3, 1]
+            axis_of_rotation = space_ops.normalize(
+                np.cross(camera_y_axis, camera_position),
+            )
+            rotation_matrix = space_ops.rotation_matrix(
+                d_point[1],
+                axis_of_rotation,
+                homogeneous=True,
+            )
+
+            maximum_polar_angle = self.camera.maximum_polar_angle
+            minimum_polar_angle = self.camera.minimum_polar_angle
+
+            potential_camera_model_matrix = rotation_matrix @ self.camera.model_matrix
+            potential_camera_location = potential_camera_model_matrix[:3, 3]
+            potential_camera_y_axis = potential_camera_model_matrix[:3, 1]
+            sign = (
+                np.sign(potential_camera_y_axis[2])
+                if potential_camera_y_axis[2] != 0
+                else 1
+            )
+            potential_polar_angle = sign * np.arccos(
+                potential_camera_location[2]
+                / np.linalg.norm(potential_camera_location),
+            )
+            if minimum_polar_angle <= potential_polar_angle <= maximum_polar_angle:
+                self.camera.model_matrix = potential_camera_model_matrix
+            else:
+                sign = np.sign(camera_y_axis[2]) if camera_y_axis[2] != 0 else 1
+                current_polar_angle = sign * np.arccos(
+                    camera_position[2] / np.linalg.norm(camera_position),
+                )
+                if potential_polar_angle > maximum_polar_angle:
+                    polar_angle_delta = maximum_polar_angle - current_polar_angle
+                else:
+                    polar_angle_delta = minimum_polar_angle - current_polar_angle
+                rotation_matrix = space_ops.rotation_matrix(
+                    polar_angle_delta,
+                    axis_of_rotation,
+                    homogeneous=True,
+                )
+                self.camera.model_matrix = rotation_matrix @ self.camera.model_matrix
+
+            # Translate to target the original target.
+            self.camera.model_matrix = (
+                opengl.translation_matrix(*self.camera_target)
+                @ self.camera.model_matrix
+            )
+        # Right click drag.
+        elif buttons == 4:
+            camera_x_axis = self.camera.model_matrix[:3, 0]
+            horizontal_shift_vector = -d_point[0] * camera_x_axis
+            vertical_shift_vector = -d_point[1] * np.cross(OUT, camera_x_axis)
+            total_shift_vector = horizontal_shift_vector + vertical_shift_vector
+
+            self.camera.model_matrix = (
+                opengl.translation_matrix(*total_shift_vector)
+                @ self.camera.model_matrix
+            )
+            self.camera_target += total_shift_vector
+
+    def set_key_function(self, char, func):
+        self.key_to_function_map[char] = func
+
+    def on_mouse_press(self, point, button, modifiers):
+        for func in self.mouse_press_callbacks:
+            func()
